refactor: part2 of signifant reorganization
