+++ .pre-commit-config.yaml
@@ -15,6 +15,11 @@ repos:
       - id: mixed-line-ending
       - id: trailing-whitespace
 
+  - repo: https://github.com/kynan/nbstripout
+    rev: 0.6.1
+    hooks:
+    - id: nbstripout
+
   # - repo: local
   #   hooks:
   #     - id: pytest
