+++ docs/_config.yml
@@ -30,8 +30,8 @@ latex:
     documentclass: book
 
 # Add a bibtex file so that we can create citations
-bibtex_bibfiles:
-  - references.bib
+# bibtex_bibfiles:
+#   - references.bib
 
 # Information about where the book exists on the web
 repository:
