+++ docs/api/atiSetup.rst
@@ -1,7 +1,7 @@
 atiSetup
 ==========================
 
-.. automodule:: atiSetup
+.. automodule:: pyamptools.atiSetup
    :members:
    :undoc-members:
    :show-inheritance:
