+++ docs/api/extract_ff.rst
@@ -1,4 +1,4 @@
 extract_ff
 ==========================
 
-.. autofunction:: extract_ff.extract_ff
+.. autofunction:: pyamptools.extract_ff.extract_ff
