--- docs/api/fit.rst
@@ -1,7 +1,0 @@
-fit
-==========================
-
-.. automodule:: fit
-   :members:
-   :undoc-members:
-   :show-inheritance:
