+++ docs/api/loadParameters.rst
@@ -1,7 +1,7 @@
 LoadParameters
 ==========================
 
-.. autoclass:: LoadParameters.LoadParameters
+.. autoclass:: pyamptools.utility.LoadParameters.LoadParameters
     :members:
     :undoc-members:
     :show-inheritance:
