+++ docs/api/mcmc.rst
@@ -3,15 +3,15 @@ MCMC
 
 MCMC
 -------
-.. automodule:: mcmc
+.. automodule:: pyamptools.mcmc
    :members:
    :undoc-members:
    :show-inheritance:
 
-MCMC/Optuna
-------------
-.. automodule:: mcmcOptimalMoves
-   :members:
-   :special-members: __init__, __call__
-   :undoc-members:
-   :show-inheritance:
+.. MCMC/Optuna
+.. ------------
+.. .. automodule:: pyamptools.mcmcOptimalMoves
+..    :members:
+..    :special-members: __init__, __call__
+..    :undoc-members:
+..    :show-inheritance:
