+++ docs/api/plotgen.rst
@@ -1,19 +1,7 @@
 PlotGen
 ==========================
 
-PlotGenRDF
-----------
-.. automodule:: PlotGenRDF
+.. automodule:: pyamptools.utility.plotgen_utils
    :members:
    :undoc-members:
    :show-inheritance:
-
-.. automodule:: plotgen_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-PlotGen
--------
-.. autofunction:: PlotGen.PlotGen
