+++ docs/demos
@@ -1,0 +1,1 @@
+../demos
