+++ set_environment.sh
@@ -43,5 +43,3 @@ else
     echo ""
     echo "ROOT not found. Please go into root and run build_root.sh to your specifications."
 fi
-
-python $REPO_HOME/utility/link_modules.py # symlink files into conda environment and root include directory
