+++ src/pyamptools/README.md
@@ -3,5 +3,3 @@
 - `extract_ff.py`: Extract fit fractions from a fit results file
 - `fit.py`: Replacement for fit.cc and fitMPI.cc
 - `mcmc.py`: Performs Markov Chain Monte
-- `PlotGen.py`: Uses PlotGen to draw fit results, replaces scripts like halld_sim/src/programs/AmplitudeAnalysis/plot_etapi0
-- `PlotGenRDF.py`: Uses RDataFrames+FSRoot to draw fit results
