+++ src/pyamptools/utility/LoadParameters.py
@@ -1,4 +1,3 @@
-# import os
 import numpy as np
 
 ############################################################
