+++ src/pyamptools/utility/RDFmacros.py
