+++ tests/gradients_test.py
@@ -15,7 +15,7 @@ def test_gradients():
     assert( os.path.isfile(cfgfile) ), f'Config file does not exist at specified path'
 
     ################### LOAD LIBRARIES ##################
-    import atiSetup
+    from pyamptools import atiSetup
     USE_MPI, USE_GPU, RANK_MPI = atiSetup.setup(globals())
 
     ############## LOAD CONFIGURATION FILE ##############
